# openapi.yaml
openapi: 3.0.3
info:
  title: LeanStock API
  version: 1.0.0
  description: |
    Production-grade inventory management backend.
    Multi-tenant, role-based, with dead-stock decay, low-stock alerts, and audit logging.
servers:
  - url: http://localhost:8000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        nextCursor: { type: string, nullable: true, example: "clxyz1234567890" }
        hasMore: { type: boolean }
...
    Product:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        sku: { type: string, example: "WIDGET-001" }
        name: { type: string, example: "Blue Widget" }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true, example: "Electronics" }
        unitPrice: { type: string, example: "49.99" }
        reorderThreshold: { type: integer, example: 10 }
        isDecayEnabled: { type: boolean }
        decayDaysThreshold: { type: integer, example: 30 }
        decayPercent: { type: string, example: "10.00" }
        createdAt: { type: string, format: date-time }
    Location:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string, example: "Main Warehouse" }
        address: { type: string, nullable: true, example: "123 Industrial Ave" }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    InventoryItem:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        productId: { type: string }
        locationId: { type: string }
        quantity: { type: integer, example: 42 }
        currentPrice: { type: string, example: "45.00" }
        originalPrice: { type: string, example: "49.99" }
        version: { type: integer, example: 3 }
        lastReceivedAt: { type: string, format: date-time }
        product:
          $ref: '#/components/schemas/Product'
        location:
          $ref: '#/components/schemas/Location'
    StockTransaction:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        productId: { type: string }
        locationId: { type: string }
        userId: { type: string }
        type:
          type: string
          enum: [INBOUND, OUTBOUND, TRANSFER_IN, TRANSFER_OUT, ADJUSTMENT]
        quantityDelta: { type: integer, example: 10 }
        note: { type: string, nullable: true }
        supplierRef: { type: string, nullable: true, example: "PO-2024-001" }
        relatedTxId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    StockAlert:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        inventoryItemId: { type: string }
        status:
          type: string
          enum: [ACTIVE, RESOLVED, SNOOZED]
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
        inventoryItem:
          $ref: '#/components/schemas/InventoryItem'
...
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Cursor from previous page (last item's id)
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
      description: Number of items per page (max 100)
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List all products (paginated)
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Product' } }
        '401': { description: Unauthorized }
    post:
      tags: [Products]
      summary: Create a product (ADMIN, MANAGER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, name, unitPrice]
              properties:
                sku: { type: string, example: "WIDGET-001" }
                name: { type: string, example: "Blue Widget" }
                description: { type: string }
                category: { type: string, example: "Electronics" }
                unitPrice: { type: number, example: 49.99 }
                reorderThreshold: { type: integer, default: 0 }
                isDecayEnabled: { type: boolean, default: false }
                decayDaysThreshold: { type: integer, default: 30 }
                decayPercent: { type: number, default: 10 }
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Product' }
        '403': { description: Forbidden }
        '409': { description: SKU already exists }
        '422': { description: Validation error }
...
  /inventory:
    get:
      tags: [Inventory]
      summary: List inventory items (paginated)
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: locationId
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/InventoryItem' } }
...
  /locations:
    get:
      tags: [Locations]
      summary: List locations (paginated)
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Location' } }
    post:
      tags: [Locations]
      summary: Create a location (ADMIN, MANAGER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Main Warehouse" }
                address: { type: string, example: "123 Industrial Ave" }
      responses:
        '201': { description: Created }
        '409': { description: Name already exists }
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions with filters (paginated)
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: productId
          in: query
          schema: { type: string }
        - name: locationId
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, enum: [INBOUND, OUTBOUND, TRANSFER_IN, TRANSFER_OUT, ADJUSTMENT] }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/StockTransaction' } }
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List stock alerts (paginated)
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, RESOLVED, SNOOZED] }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/StockAlert' } }
  /alerts/{id}:
    patch:
      tags: [Alerts]
      summary: Update alert status (ADMIN, MANAGER)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [RESOLVED, SNOOZED] }
      responses:
        '200': { description: Updated }
        '404': { description: Alert not found }

  # ─── Reports ──────────────────────────────────────────────────────────
...